Common: Move `infection_monkey.utils.decorators` to common
